Merged PR 52495: Email capture for Hyvä Checkout
## What's being changed

This PR adds a new compatibility module for Hyvä Checkout. It leans on both the main compatibility module `Hyva_DotdigitalgroupEmail`, and of course the parent module `Dotdigitalgroup_Email`.

## Why it's being changed

To support email capture in Hyvä Checkout. In addition to calling `window.dmPt('identify', email);`, the checkout flavour of email capture makes an AJAX request to a route in `Dotdigitalgroup_Email` to add an email address to the active guest cart, enabling abandoned carts for guests.

## How to review / test this change

- Enable Hyvä Checkout for your Hyvä-themed website
- Test email capture for a guest cart, confirming that the quote table is updated
- Confirm that Luma Checkout still works as expected in a parallel store view (under a Luma website).

Related work items: #242867